factory-py-movie-master/server/requirements.txt
asgiref==3.4.1 #type: ignore
autopep8==1.6.0 #type: ignore
Django==4.0 #type: ignore
django-cors-headers==3.10.1 #type: ignore
django-storages==1.12.3 #type: ignore
djangorestframework==3.13.1 #type: ignore
djangorestframework-simplejwt==5.0.0 #type: ignore
Pillow==10.0.0 #type: ignore
psycopg2-binary==2.9.9 #type: ignore
pycodestyle==2.8.0 #type: ignore
PyJWT==2.3.0 #type: ignore
pytz==2021.3 #type: ignore
sqlparse==0.4.2 #type: ignore
toml==0.10.2 #type: ignore
tzdata==2021.5 #type: ignore
whitenoise==5.3.0 #type: ignore